% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RoxygenTools.R
\name{add_importFrom_statements}
\alias{add_importFrom_statements}
\title{Main Function to Process R Script for Package Calls}
\usage{
add_importFrom_statements(
  file_path,
  suffix = "ADDED_BY_add_importFrom_statements",
  exclude_packages = c("MarkdownReports"),
  sure = FALSE
)
}
\arguments{
\item{file_path}{The path to the R script file.}

\item{suffix}{A string suffix to add to the \verb{@importFrom} lines.
Default: "ADDED_BY_add_importFrom_statements".}

\item{exclude_packages}{Packages to exclude from adding \verb{@importFrom} statements.
Default: c("MarkdownReports").}

\item{sure}{Logical indicating whether to proceed with the function.}
}
\description{
Reads an R script file, processes its content to find and add \verb{@importFrom} statements
for package function calls in function bodies. The statements are added to the Roxygen documentation
blocks of the functions. Default: Excludes "MarkdownReports" from processing.
}
