% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PackageTools.R
\name{all_funs}
\alias{all_funs}
\title{List All Functions in a Package}
\usage{
all_funs(packageName)
}
\arguments{
\item{packageName}{A character string specifying the name of the package.}
}
\value{
A character vector containing the names of all functions in the specified package,
excluding internal functions and certain predefined objects.
}
\description{
Lists all function names available in a specified R package. It excludes certain
internal objects and functions that are not intended for end users.
}
\note{
The function excludes objects like ".\strong{DEVTOOLS}", ".\strong{NAMESPACE}.", and
".\strong{S3MethodsTable}." Additionally, it does not list functions that start with a dot ('.'),
which are usually internal.
}
\examples{
all_funs("dplyr") # Lists all functions in the 'dplyr' package

}
