% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DependencyTools.R
\name{analyze_function_dependencies}
\alias{analyze_function_dependencies}
\title{Analyze Function Dependencies with Exclusions}
\usage{
analyze_function_dependencies(
  func_name,
  package_name,
  ls_fun_names_to_map,
  exclude_packages = c("base", "utils", "methods", "stats"),
  exclude_strings = c("HYPERLINK", "Deprecated")
)
}
\arguments{
\item{func_name}{The name of the function to analyze.
Default: None, a valid function name must be provided.}

\item{package_name}{The name of the package containing the function.
Default: None, a valid package name must be provided.}

\item{ls_fun_names_to_map}{A list of function names mapped to their source package.
Default: None, a valid mapping must be provided.}

\item{exclude_packages}{A vector of package names whose functions should be excluded from the analysis.
Default: c("base", "utils", "methods", "stats").}

\item{exclude_strings}{A vector of strings to exclude from the dependencies.
Default: c("HYPERLINK", "Deprecated")}
}
\value{
A vector of function names that the specified function depends on, excluding specified strings.
}
\description{
Analyzes dependencies of functions within a package, excluding specified packages and strings.
}
\examples{
analyze_function_dependencies(
  func_name = "column_to_rownames", package_name = "tibble",
  ls_fun_names_to_map = map_functions_to_packages("tibble"),
  exclude_packages = c("base", "utils", "methods", "stats"),
  exclude_strings = c("HYPERLINK", "Deprecated")
)

}
