% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DependencyTools.R
\name{analyze_package_dependencies}
\alias{analyze_package_dependencies}
\title{Analyze Package Dependencies}
\usage{
analyze_package_dependencies(
  packages,
  exclude_packages = c("base", "utils", "methods", "stats"),
  extended_search = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{packages}{A vector of package names to analyze.
Default: None, a valid vector of package names must be provided.}

\item{exclude_packages}{A vector of package names whose functions should be excluded from the analysis.
Default: c("base", "utils", "methods", "stats").}

\item{extended_search}{Logical; whether to perform an extended search across all installed packages.
Default: FALSE.}

\item{verbose}{Print logs? Default: FALSE.}
}
\value{
A list of dependencies for each function in the specified packages.
}
\description{
Analyzes and maps the dependencies of all functions in a set of packages.
}
\examples{
analyze_package_dependencies(c("stats", "utils"))
}
