% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PackageTools.R
\name{checkGlobalVarsInPackage}
\alias{checkGlobalVarsInPackage}
\title{Check for Global Variables in Package Functions}
\usage{
checkGlobalVarsInPackage(packageName, warn = TRUE)
}
\arguments{
\item{packageName}{A string specifying the name of the package.}

\item{warn}{A logical value indicating whether to display warnings for global variable usage.
Default is TRUE.}
}
\description{
\code{checkGlobalVarsInPackage} iterates over all functions in a specified package
and checks each function for the usage of global variables using \code{isoENV::checkGlobalVars}.
}
\details{
The function uses \code{all_funs()} to retrieve all function names from the specified package.
Then, it checks for global variable usage in each function. Warnings can be toggled on or off.
}
\note{
This function depends on \code{isoENV::checkGlobalVars} to check for global variable usage.
Ensure that the isoENV package is installed and loaded before using this function.
}
\examples{
# Assuming the isoENV package is loaded and all_funs() is defined
checkGlobalVarsInPackage("dplyr")

}
