% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Miscellaneous.R
\name{checkScriptEnv_v1}
\alias{checkScriptEnv_v1}
\title{Check Script Environment}
\usage{
checkScriptEnv_v1(
  path,
  input.variables,
  exclude_var = c("i", "path"),
  env = NULL,
  packages_load = c("Seurat"),
  packages_load_default = c("utils", "grDevices", "graphics", "stats", "methods",
    "ggplot2", "Stringendo", "ReadWriter", "CodeAndRoll2", "MarkdownHelpers",
    "MarkdownReports", "ggExpress", "Seurat.utils", "isoENV", "UVI.tools",
    "Connectome.tools", "NestedMultiplexer"),
  replace_missing_calls = FALSE
)
}
\arguments{
\item{path}{Path to the R script file.}

\item{input.variables}{A character vector of variable names to be copied to the new environment.}

\item{exclude_var}{A character vector of variable names to be excluded from the report.}

\item{env}{Optional; an environment to use instead of creating a new one.}

\item{packages_load}{A character vector of package names to load functions from.}

\item{packages_load_default}{A character vector of default package names to load functions from.}

\item{replace_missing_calls}{Logical; if TRUE, replaces missing function calls in the script with their fully qualified names.}
}
\description{
Checks if all functions and variables called in a script are found in a specified environment.
Optionally replaces missing function calls in the script with their fully qualified names.
}
