% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Miscellaneous.R
\name{copy_github_badge}
\alias{copy_github_badge}
\title{Copy GitHub Badge Markdown Image Link to Clipboard}
\usage{
copy_github_badge(
  status = "experimental",
  prefix = "https://raw.githubusercontent.com/vertesy/TheCorvinas/master/GitHub/Badges/",
  copy_to_clipboard = FALSE
)
}
\arguments{
\item{status}{The status of the project, determining which badge is copied.
Valid options are 'experimental', 'active', 'archive', and 'hibernate'.
Default: 'experimental'.}

\item{prefix}{The URL prefix for the badge images.
Default: "https://raw.githubusercontent.com/vertesy/TheCorvinas/master/GitHub/Badges/".}

\item{copy_to_clipboard}{Logical. If TRUE, the badge Markdown code is copied to the clipboard.}
}
\value{
The function does not return a value but copies the relevant badge Markdown code to the clipboard.
}
\description{
This function copies the Markdown code for a GitHub badge to the clipboard based on the
specified status. It supports four statuses: 'experimental', 'active', 'archive', and
'hibernate'.
}
\examples{
copy_github_badge("experimental")
copy_github_badge("active", "https://example.com/badges/")
}
