% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DocumentationTools.R
\name{document_and_create_package}
\alias{document_and_create_package}
\title{Create R Package from Configuration}
\usage{
document_and_create_package(
  package_dir,
  config_file = "config.R",
  backup_r_script = FALSE,
  update_citation = TRUE,
  dev_folder = "Development"
)
}
\arguments{
\item{package_dir}{The path to the package directory. Example: '~/GitHub/Packages/PackageX'.}

\item{config_file}{The configuration file name within the package's Development directory.
Default: 'config.R'.}

\item{backup_r_script}{Whether to back up the previous R script into another file. Default: FALSE.}

\item{update_citation}{Whether to update the CITATION file. Default: FALSE.}

\item{dev_folder}{The name of the development folder. Default: 'Development'.}
}
\value{
None
}
\description{
Automates the creation of an R package from a configuration file.
This function automates the creation of an R package by sourcing a configuration file
from the specified package directory. It assumes the presence of a \code{config.R} file in
the \code{Development} subdirectory of the package.
}
\examples{
document_and_create_package("~/GitHub/Packages/PackageX", "config.R", TRUE)

}
