% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PackageTools.R
\name{.convertFilePathToOutput}
\alias{.convertFilePathToOutput}
\title{Convert File Path for Documentation}
\usage{
.convertFilePathToOutput(
  inputPath,
  fn_prefix = "list.of.functions.in",
  ext = ".md"
)
}
\arguments{
\item{inputPath}{The path to the R script file.
Default: None, a valid file path must be provided.}

\item{fn_prefix}{A prefix added before the script's base name in the output markdown filename.
Default: "list.of.functions.in.".}

\item{ext}{The file extension for the output file.
Default: ".md".}
}
\value{
A string representing the converted file path in the specified markdown format.
}
\description{
Converts a file path from an R script format to a markdown file format,
indicating that it contains a documentation list, such as a list of functions.
The function allows specifying a prefix for the markdown filename and the file extension.
}
\examples{
\dontrun{
}
}
