% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DependencyTools.R
\name{.count_and_print_function_summary}
\alias{.count_and_print_function_summary}
\title{Count and Print the Number of Functions in Results}
\usage{
.count_and_print_function_summary(dependencies)
}
\arguments{
\item{dependencies}{Results from the above functions' dependencies}
}
\value{
Nothing.
}
\description{
Private function to count and print the number of functions in results ("dependencies").
}
