% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PackageTools.R
\name{.get_description_from_config}
\alias{.get_description_from_config}
\title{Helper to Get Description from Config}
\usage{
.get_description_from_config(file)
}
\arguments{
\item{file}{A character string specifying the path to the current script or file of interest.
The function constructs the path to \code{config.R} based on this path. Default: "".}
}
\value{
Returns the "description" field from the \code{DESCRIPTION} object if the conditions are met.
Returns \code{NULL} otherwise.
}
\description{
This function retrieves the "description" field from the \code{DESCRIPTION} object,
which is sourced from a \code{config.R} file located two directories above the specified file. The
function checks for the existence of \code{config.R}, sources it, and validates the presence and
non-nullity of the "description" field in the \code{DESCRIPTION} object.
}
