% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DocumentationTools.R
\name{extract_package_dependencies}
\alias{extract_package_dependencies}
\title{Extract Package Dependencies}
\usage{
extract_package_dependencies(
  package_dir,
  output_file = "Development/Dependencies.R"
)
}
\arguments{
\item{package_dir}{The path to the package directory. Default: '~/GitHub/Packages/PackageX'.}

\item{output_file}{The relative path from the package directory to the dependencies file.
Default: 'Development/Dependencies.R'.}
}
\value{
None
}
\description{
This function checks the package dependencies by listing functions used in all R scripts
found in a specified package directory and writes them to a dependencies file, separated
by script.
}
\examples{
extract_package_dependencies(
  "~/GitHub/Packages/PackageX",
  "Development/Dependencies.R"
)

}
