% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DependencyTools.R
\name{filter_dependencies}
\alias{filter_dependencies}
\title{Filter Function Dependencies}
\usage{
filter_dependencies(dependencies, include_only_with_deps = TRUE)
}
\arguments{
\item{dependencies}{A list of dependencies for each function.
Default: None, a valid list of dependencies must be provided.}

\item{include_only_with_deps}{Logical; if TRUE, only includes functions with dependencies.
Default: TRUE.}
}
\value{
A filtered list of dependencies.
}
\description{
Filters the function dependencies to either include only functions with dependencies
or only those without dependencies.
}
\examples{
deps <- analyze_package_dependencies(c("stats", "utils"))
filter_dependencies(deps)
}
