% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RoxygenTools.R
\name{find_package_calls}
\alias{find_package_calls}
\title{Finding \code{::} Usage within Function Bodies}
\usage{
find_package_calls(
  body_content,
  pattern = "\\\\b\\\\w+::[\\\\w.]+\\\\b",
  exclude_packages = c()
)
}
\arguments{
\item{body_content}{A character vector of lines in a function body.}

\item{pattern}{A regular expression pattern to identify package function calls.
Default: "\\b\\w+::\\w+\\b".}

\item{exclude_packages}{A character vector of package names to exclude from the search.
Default: empty character vector.}
}
\value{
A unique list of package calls found in the body content.
}
\description{
Searches for package function calls using the \code{::} operator within the given content of
function bodies, excluding specified packages. Default: Searches for \code{::} usage, excluding packages
listed in \code{exclude_packages}.
}
