% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Miscellaneous.R
\name{find_package_of_this_function}
\alias{find_package_of_this_function}
\title{Find which Package a given Function belongs to}
\usage{
find_package_of_this_function(
  functionName,
  searchInstalled = FALSE,
  toClipboard = TRUE
)
}
\arguments{
\item{functionName}{The name of the function (character or function object).}

\item{searchInstalled}{If TRUE, searches all installed packages (default FALSE).}

\item{toClipboard}{If TRUE, copies the result to the clipboard (default TRUE).}
}
\value{
The name of the package containing the function, or NULL if not found.
}
\description{
Determines the package that a given function is defined in and optionally copies
the result to the clipboard. Warns if the function is found in multiple packages.
}
\examples{
 find_package_of_this_function('sppp')
 find_package_of_this_function(sppp)
 find_package_of_this_function(filter)
 find_package_of_this_function("filterzzz")

}
