% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RoxygenTools.R
\name{get_function_bodies}
\alias{get_function_bodies}
\title{Extract Function Bodies from R Script}
\usage{
get_function_bodies(file_content)
}
\arguments{
\item{file_content}{A character vector representing the lines of an R script.}
}
\value{
A list where each element is a vector of line numbers representing a function body.
}
\description{
This function identifies the start and end lines of each function body in an R script.
Each function body is extracted for further processing. Default: Extracts function bodies from
a given R script content.
}
