% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DependencyTools.R
\name{get_package_functions}
\alias{get_package_functions}
\title{Get Package Functions}
\usage{
get_package_functions(package_name)
}
\arguments{
\item{package_name}{The name of the package.
Default: None, a valid package name must be provided.}
}
\value{
A character vector of function names from the package.
}
\description{
Retrieves all function names from a specified package.
}
\examples{
get_package_functions("stats")
}
