% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DependencyTools.R
\name{get_parse_pkg_deps}
\alias{get_parse_pkg_deps}
\title{Parse and Get Package Dependencies for an R Package}
\usage{
get_parse_pkg_deps(pkg = rstudioapi::getActiveProject())
}
\arguments{
\item{pkg}{Directory containing R scripts from which to extract dependencies.
Default: Active RStudio project's 'R' directory (assumes an RStudio project environment).}
}
\value{
A character vector of package dependencies found in the R scripts.
}
\description{
Extracts package dependencies from R scripts in a specified directory, typically
the 'R/' directory of an R package. This function leverages 'renv::dependencies()'
to parse the scripts and extract package names. It then appends these dependencies
to a specified configuration file.
}
\examples{
get_parse_pkg_deps() # Assuming an RStudio project
}
