% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PackageTools.R
\name{list_of_funs_to_markdown}
\alias{list_of_funs_to_markdown}
\title{Parse Roxygen Comments from R Script}
\usage{
list_of_funs_to_markdown(
  file,
  output_file = .convertFilePathToOutput(file, ext = ".det.md"),
  write_title_field = TRUE,
  fun_header_level = "####",
  open_results = TRUE
)
}
\arguments{
\item{file}{The path to the R script file to be parsed.
Default: None, a valid file path must be provided.}

\item{output_file}{The path to the output markdown file where the summary will be written.
Default: Generated using .convertFilePathToOutput() with ".det.md" extension.}

\item{write_title_field}{Logical; whether to include the title field in the output.
Default: TRUE.}

\item{fun_header_level}{The markdown header level to be used for function names.
Default: "####".}

\item{open_results}{Open resulting file? Default: TRUE.}
}
\value{
This function does not return a value; it writes output to the specified markdown file.
}
\description{
This function parses a given R script for Roxygen comments, extracts function titles and descriptions,
and writes a summary to an output markdown file. The output file can have a custom name, and the
function allows specifying the header level for functions in the markdown file.
}
\examples{
\dontrun{
list_of_funs_to_markdown("path/to/your_script.R", "path/to/output_file.md")
}
}
