% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PackageTools.R
\name{list_of_funs_to_markdown_simple}
\alias{list_of_funs_to_markdown_simple}
\title{Parse Roxygen Comments}
\usage{
list_of_funs_to_markdown_simple(
  file,
  output_file = .convertFilePathToOutput(file),
  fun_header_level = "####",
  open_results = TRUE
)
}
\arguments{
\item{file}{Path to the R script file to be parsed.
Default: None, a valid file path must be provided.}

\item{output_file}{Path to the output file where the summary will be written.
Default: "1list.of.functions.in.YOURFILE.md"}

\item{fun_header_level}{Header level for functions. Default: "####"}

\item{open_results}{Open resulting file? Default: TRUE.}
}
\value{
This function does not return a value; it writes output to the specified file.
}
\description{
Extracts and summarizes Roxygen documentation comments from a specified R script file.
This function reads an R script, identifies Roxygen comments for function titles and descriptions,
and writes a summary to an output file.
}
\examples{
\dontrun{
list_of_funs_to_markdown_simple("path/to/your_script.R", "path/to/output_file.txt")
}
}
