% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DependencyTools.R
\name{map_functions_to_packages}
\alias{map_functions_to_packages}
\title{Map Functions to Packages}
\usage{
map_functions_to_packages(
  packages,
  extended_search = FALSE,
  verbose = TRUE,
  exclude_from_extended = character()
)
}
\arguments{
\item{packages}{A vector of package names to include in the mapping.
Default: None, a valid vector of package names must be provided.}

\item{extended_search}{Logical; whether to perform an extended search across all installed packages.
Default: FALSE.}

\item{verbose}{Print logs? Default: TRUE.}

\item{exclude_from_extended}{A vector of package names to exclude from the extended search.
Default: Empty character vector.}
}
\value{
A list mapping function names to their source package.
}
\description{
Creates a mapping of functions to their source packages.
Optionally performs an extended search across all installed packages.
}
\examples{
map_functions_to_packages(c("stats", "utils"), extended_search = TRUE)
}
