% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Miscellaneous.R
\name{openReadme}
\alias{openReadme}
\title{Open README.md of the Current Active RStudio Project}
\usage{
openReadme()
}
\value{
Invisible NULL. The function is used for its side effect of opening a file.
}
\description{
Opens the README.md file located in the root of the currently active RStudio project.
This function automatically detects the operating system to use the appropriate
system command for opening the file.
}
\examples{
openReadme() # Opens README.md in the active RStudio project
}
