% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DocumentationTools.R
\name{parse_description}
\alias{parse_description}
\title{Parse DESCRIPTION File}
\usage{
parse_description(config_path)
}
\arguments{
\item{config_path}{The file path of the configuration file.}
}
\value{
A list representing the DESCRIPTION file.
}
\description{
Helper function to parse the DESCRIPTION file from a configuration file.
}
