% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PackageTools.R
\name{parse_rmd_vignette_from_roxygen}
\alias{parse_rmd_vignette_from_roxygen}
\title{Parse RMD Vignette from Roxygen}
\usage{
parse_rmd_vignette_from_roxygen(
  file,
  output_file = .convertFilePathToOutput(file, fn_prefix = "Vignette", ext = ".Rmd"),
  fun_header_level = "####",
  open_results = TRUE,
  parse_examples = TRUE,
  package_desc = .get_description_from_config(file)
)
}
\arguments{
\item{file}{The path to the R script file to be parsed. Default: "".}

\item{output_file}{Path to the output file where the summary will be written. Default is generated
using \code{.convertFilePathToOutput} with prefix "Vignette" and extension ".Rmd".}

\item{fun_header_level}{The markdown header level for function names. Default: "####".}

\item{open_results}{If TRUE, attempts to open the resulting file automatically. Default: TRUE.}

\item{parse_examples}{If TRUE, function examples are parsed and included in the vignette.
Default: TRUE.}

\item{package_desc}{The package description, automatically retrieved from configuration.
Default is obtained via \code{.get_description_from_config(file)}.}
}
\value{
Does not return a value; it writes output to the specified file.
}
\description{
Extracts and summarizes Roxygen documentation comments from a specified R script file
and creates a simple R Markdown vignette. It reads an R script, identifies Roxygen comments for
function titles and descriptions, and writes a summary to an output file. If \code{parse_examples} is
TRUE, it also includes examples found in the Roxygen comments.
}
\examples{
parse_rmd_vignette_from_roxygen("path/to/script.R")

}
