% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReplacementTools.R
\name{replace_a_string_in_a_file}
\alias{replace_a_string_in_a_file}
\title{Replace a string in a file with options for whole word and case matching}
\usage{
replace_a_string_in_a_file(
  file_path,
  from,
  to,
  whole_word = TRUE,
  ignore_case = FALSE,
  perl = TRUE,
  backup = FALSE,
  ...
)
}
\arguments{
\item{file_path}{The path to the file where replacements will be made. Default: none.}

\item{from}{The string to be replaced. Default: none.}

\item{to}{The string that will replace \code{from}. Default: none.}

\item{whole_word}{Logical. If \code{TRUE}, replaces only whole word matches. Default: \code{TRUE}.}

\item{ignore_case}{Logical. If \code{TRUE}, the replacement is case insensitive. Default: \code{FALSE}.}

\item{perl}{Logical. If \code{TRUE}, use Perl-compatible regular expressions. Default: \code{TRUE}.}

\item{backup}{Logical. If \code{TRUE}, a backup of the original file is saved with a \code{.bac} extension.
Default: \code{FALSE}.}

\item{...}{Additional arguments to be passed to \code{gsub()} and \code{grepl()}.}
}
\value{
The function performs replacements in the file, returning the number of instances replaced
and outputting a message indicating the completion of the replacement process.
}
\description{
The \code{replace_a_string_in_a_file()} function replaces all instances of a string in a file, with
options to replace only whole word matches and to replace with case sensitivity or ignore case.
It also allows backing up the file before making the replacements.
}
\examples{
# Replace "from" with "to" in "my_file.txt", with case matching and whole word match.
replace_a_string_in_a_file("my_file.txt", "from", "to")

# Replace "from" with "to" without matching the case and ignoring whole word match.
replace_a_string_in_a_file("my_file.txt", "from", "to", whole_word = FALSE, ignore_case = FALSE)

}
