% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PackageSetupTools.R
\name{setup_new_package_from_template}
\alias{setup_new_package_from_template}
\title{Setup New Package from Template}
\usage{
setup_new_package_from_template(
  package_name = "NewPackage",
  packages_path = "~/GitHub/Packages/",
  template_pkg = "PackageTools",
  place_holder = "TEMPLATE"
)
}
\arguments{
\item{package_name}{Name of the new package. Default: 'NewPackage'.}

\item{packages_path}{Path to the packages directory. Default: '~/GitHub/Packages/'.}

\item{template_pkg}{Name of the template package. Default: 'PackageTools'.}

\item{place_holder}{Placeholder string to replace. Default: 'TEMPLATE'.}
}
\value{
None
}
\description{
Setup a new R package by copying template contents and replacing placeholders.
}
\details{
Setup New Package from Template

This function sets up a new R package by copying contents from a template directory,
replacing template placeholders in file names and file contents with the specified package name.
}
\examples{
setup_new_package_from_template("MyNewPackage", "~/GitHub/Packages/", "PackageTools")
}
