% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DocumentationTools.R
\name{update_citation_file}
\alias{update_citation_file}
\title{Helper function to update the CITATION file of a package.}
\usage{
update_citation_file(RepositoryDir, version)
}
\arguments{
\item{RepositoryDir}{The directory of the repository.}

\item{version}{The version of the package.}
}
\value{
None
}
\description{
Update the CITATION.cff file of a package based on its version.
}
