% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RoxygenTools.R
\name{add_import_from}
\alias{add_import_from}
\title{Add 'importFrom' Statements to Roxygen Blocks}
\usage{
add_import_from(
  file_content,
  bodies,
  suffix = "ADDED_BY_add_importFrom_statements",
  exclude_packages = c()
)
}
\arguments{
\item{file_content}{A character vector representing the lines of an R script.}

\item{bodies}{A list of function body line numbers.}

\item{suffix}{A string suffix to add to the\verb{ @importFrom} lines.
Default: "ADDED_BY_add_importFrom_statements".}

\item{exclude_packages}{Packages to exclude from adding \verb{@importFrom} statements.
Default: empty character vector.}
}
\value{
Modified file content with added \verb{@importFrom} statements.
}
\description{
Iterates over function bodies in an R script to identify package function calls and
adds corresponding \verb{@importFrom} statements to the Roxygen blocks of these functions.
Default: Suffix for added lines is "ADDED_BY_add_importFrom_statements".
}
