% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReplacementTools.R
\name{.safely_replace_calls}
\alias{.safely_replace_calls}
\title{Safely Replace Short Function Calls in a Line of R Script}
\usage{
.safely_replace_calls(line, strict_mode)
}
\arguments{
\item{line}{A single line from an R script.}

\item{strict_mode}{A boolean flag to determine the strictness of the match.
If \code{TRUE}, matches are made only when not part of larger alphanumeric strings.
If \code{FALSE}, all instances are replaced.}
}
\value{
A string representing the modified line.
}
\description{
Safely replaces instances of \verb{length(} with \verb{length(} and \code{p0} with \verb{paste0(} in a given line of R script.
Operates in strict mode to ensure that replacements are made only when not part of a larger word or variable name.
}
