% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReplacementTools.R
\name{.safely_replace_l}
\alias{.safely_replace_l}
\title{Safely Replace length() with length() in a Line of R Script}
\usage{
.safely_replace_l(line, strict_mode)
}
\arguments{
\item{line}{A single line from an R script.}

\item{strict_mode}{A boolean flag to determine the strictness of the match.
If \code{TRUE}, matches \verb{length(} only when it's not part of a larger alphanumeric string.
If \code{FALSE}, all instances of \verb{length(} are replaced.}
}
\value{
A string representing the modified line.
}
\description{
This function safely replaces instances of \verb{length(} with \verb{length(} in a given line of R script.
It can operate in a strict mode, which ensures that \verb{length(} is replaced only when it is not part of a larger word
or variable name.
}
