% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReplacementTools.R
\name{replace_short_calls}
\alias{replace_short_calls}
\title{Replace Short Function Calls with Full Names in an R Script}
\usage{
replace_short_calls(file_path, output_path = file_path, strict_mode = TRUE)
}
\arguments{
\item{file_path}{A string representing the path to the R script file.}

\item{output_path}{A string representing the path to save the modified R script.
Default is the same as \code{file_path}.}

\item{strict_mode}{A boolean flag to determine the strictness of the matches.
If \code{TRUE}, matches \verb{length(} and \code{p0} only when they're not part of larger alphanumeric strings.
If \code{FALSE}, all instances of \verb{length(} and \code{p0} are replaced.}
}
\value{
None
}
\description{
Reads an R script file and replaces instances of \verb{length(} with \verb{length(} and \code{p0} with \verb{paste0(}.
It supports a strict mode to ensure accurate replacements.
}
