% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PackageTools.R
\name{source_file_stats_analyzer}
\alias{source_file_stats_analyzer}
\title{Analyze File for Code and Comment Statistics}
\usage{
source_file_stats_analyzer(
  file_path,
  pattern = "^\\\\s*#",
  patter_sourced_files = "source\\\\s*\\\\(\\\\s*['\\"]([^'\\"]+)['\\"]\\\\s*\\\\)"
)
}
\arguments{
\item{file_path}{The path to the file to be analyzed. Default: None (mandatory).}

\item{pattern}{The regular expression pattern used to identify comment lines. Default: "^\\s*#".}

\item{patter_sourced_files}{The regular expression pattern used to identify lines where files are sourced.
Default: "source\\s*\\(\\s*\href{[^'\"]+}{'\"}\link[='\\"]{'\"}\\s*\\)".}
}
\value{
A list containing the number of lines of code, the number of comment lines,
and the names of any sourced files.
}
\description{
This function analyzes a given file, counting the number of lines of code and comments.
It also identifies files that are sourced within the provided file. The function uses regular
expressions to differentiate between code and comment lines and to extract the names of sourced files.
}
\examples{
# Example usage:
source_file_stats_analyzer("path/to/your/script.R")
}
